requirements.txt
attrs==19.3.0
aws-cdk.assets==1.56.0
aws-cdk.aws-autoscaling==1.56.0
aws-cdk.aws-autoscaling-common==1.56.0
aws-cdk.aws-certificatemanager==1.56.0
aws-cdk.aws-cloudformation==1.56.0
aws-cdk.aws-cloudwatch==1.56.0
aws-cdk.aws-codeguruprofiler==1.56.0
aws-cdk.aws-ec2==1.56.0
aws-cdk.aws-efs==1.56.0
aws-cdk.aws-elasticloadbalancing==1.56.0
aws-cdk.aws-elasticloadbalancingv2==1.56.0
aws-cdk.aws-events==1.56.0
aws-cdk.aws-iam==1.56.0
aws-cdk.aws-kms==1.56.0
aws-cdk.aws-lambda==1.56.0
aws-cdk.aws-logs==1.56.0
aws-cdk.aws-rds==1.56.0
aws-cdk.aws-route53==1.56.0
aws-cdk.aws-s3==1.56.0
aws-cdk.aws-s3-assets==1.56.0
aws-cdk.aws-sam==1.56.0
aws-cdk.aws-secretsmanager==1.56.0
aws-cdk.aws-sns==1.56.0
aws-cdk.aws-sqs==1.56.0
aws-cdk.aws-ssm==1.56.0
aws-cdk.cloud-assembly-schema==1.56.0
aws-cdk.core==1.56.0
aws-cdk.custom-resources==1.56.0
aws-cdk.cx-api==1.56.0
aws-cdk.region-info==1.56.0
bc-python-hcl2==0.3.5
boto3==1.12.43
botocore==1.15.49
bridgecrew==1.0.465
cattrs==1.0.0
certifi==2020.6.20
chardet==3.0.4
checkov==1.0.465
colorama==0.4.3
constructs==3.0.4
docopt==0.6.2
docutils==0.15.2
dpath==1.5.0
gitdb==4.0.5
GitPython==3.1.7
idna==2.8
jmespath==0.10.0
jsii==1.9.0
junit-xml==1.8
lark-parser==0.7.8
publication==0.0.3
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.22.0
s3transfer==0.3.3
six==1.13.0
smmap==3.0.4
tabulate==0.8.6
termcolor==1.1.0
typing-extensions==3.7.4.2
urllib3==1.25.7
